Initial support for statement blocks in if conditions
The tokenizer now generates whitespace tokens for line beginnings to keep
information about the line indentation. The parser uses this information
to keep track of statement blocks, including nested ones.

The code is somewhat generic and should work for function bodies and other
blocks too, but for now it does not support definition of scopes.